refactor: user and auth folders
