+++ src/helpers/interfaces/authenticated-user.interface.ts
