+++ src/infra/auth.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { AuthService } from './services/auth.service';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule, JwtService } from '@nestjs/jwt';
-import { JwtStrategy } from '../infra/common/strategies/jwt.estrategy';
+import { JwtStrategy } from './common/strategies/jwt.estrategy';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from '@/users/entities/user.entity';
-import { AuthRepository } from '../users/repositories/auth.repository';
-import { AuthController } from '../infra/http/auth/auth.controller';
+import { UserEntity } from '@/infra/entities/user.entity';
+import { AuthRepository } from './repositories/auth.repository';
+import { AuthController } from './http/auth/auth.controller';
 @Module({
   imports: [
     TypeOrmModule.forFeature([UserEntity]),
