+++ src/infra/common/strategies/jwt.estrategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { AuthService } from '../../../auth/auth.service';
-import { UserEntity } from '@/users/entities/user.entity';
-import { JwtPayload } from '../../../auth/interfaces';
+import { AuthService } from '../../services/auth.service';
+import { UserEntity } from '@/infra/entities/user.entity';
+import { JwtPayload } from '../../../helpers/interfaces';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
