+++ src/infra/config/typeorm/orm-cli-config.ts
@@ -1,6 +1,6 @@
 import { DataSource, DataSourceOptions } from 'typeorm';
 import { MovieEntity } from '../../entities/movie.entity';
-import { UserEntity } from '@/users/entities/user.entity';
+import { UserEntity } from '@/infra/entities/user.entity';
 
 export const dataSourceOptions: DataSourceOptions = {
   type: 'postgres',
