+++ src/infra/repositories/auth.repository.ts
@@ -2,11 +2,11 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { NotFoundException } from '@nestjs/common';
-import { SigninDto } from '../../infra/http/auth/dto/sign-in.dto';
+import { SigninDto } from '../http/auth/dto/sign-in.dto';
 import { SignupDto } from '@/infra/http/auth/dto/sign-up.dto';
 import { encrypt } from '@/helpers/encrypter/encrypter';
-import { AuthenticatedUser } from '@/auth/interfaces';
-import { UserEntity } from '@/users/entities/user.entity';
+import { AuthenticatedUser } from '@/helpers/interfaces';
+import { UserEntity } from '@/infra/entities/user.entity';
 import { EmailIsTakenError } from '@/helpers/errors';
 import { JwtService } from '@nestjs/jwt';
 import { JwtPayload } from 'jsonwebtoken';
