+++ src/infra/users.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
+import { UsersService } from './services/users.service';
+import { UsersController } from './http/users/users.controller';
 import { UsersRepository } from './repositories/users.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from './entities/user.entity';
-import { AuthModule } from '@/auth/auth.module';
+import { AuthModule } from '@/infra/auth.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity]), AuthModule],
